Create new MOST header
