+++ Source/BoundaryConditions/Make.package
@@ -24,4 +24,5 @@ CEXE_headers += ERF_TimeInterpolatedData.H
 CEXE_headers += ERF_MOSTAverage.H
 CEXE_sources += ERF_MOSTAverage.cpp
 
+CEXE_headers += ERF_MOST.H
 CEXE_headers += ERF_MOSTStress.H
